Add settings for interactive objects in builder submenu

+++ Assets/Scripts/TrackBuilder/Battery.cs
@@ -6,7 +6,7 @@ namespace Builder
 {
     public class Battery : MonoBehaviour
     {
-        public int energy;
+        public float energy;
 
         private void OnTriggerEnter(Collider other)
         {
